Split the log in main.c onto program.c to make it testable
